Some model classes refactoring
