+++ src/Memores.Metrics.Wcf/Model/Reports/ApdexReport.cs
@@ -1,0 +1,7 @@
+﻿using Memores.Metrics.Wcf.Model.Reports.Base;
+
+namespace Memores.Metrics.Wcf.Model.Reports {
+    public class ApdexReport: MetricsReportBase {
+        public long Apdex { get; set; }
+    }
+}
