+++ src/Memores.Metrics.Wcf/Model/Reports/Base/MetricsReportBase.cs
@@ -1,34 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Memores.Metrics.Wcf.Model {
-    public class MetricsReport {
+namespace Memores.Metrics.Wcf.Model.Reports.Base {
+    public class MetricsReportBase {
 
-        public MetricsReport() {
+        public MetricsReportBase() {
             TimeStamp = DateTime.Now;
             DateStart = DateTime.Now;
         }
 
         public MetricsReportTypes MetricsReportType { get; set; }
-
-        public long Rate1m { get; set; }
-
-        public long Rate5m { get; set; }
-
-        public long Rate15m { get; set; }
-
-        public long Apdex { get; set; }
-
-        public string OperationName { get; set; }
-
+        
         public DateTime TimeStamp { get; set; }
 
         public DateTime DateStart { get; set; }
 
         public DateTime DateEnd { get; set; }
-
-        public long ProcessingTime { get; set; }
-
+        
         public List<Tag> Tags { get; set; }    
     }
 }
