+++ src/Memores.Metrics.Wcf/Model/Reports/ServiceCallReport.cs
@@ -1,0 +1,5 @@
+﻿using Memores.Metrics.Wcf.Model.Reports.Base;
+
+namespace Memores.Metrics.Wcf.Model.Reports {
+    public class ServiceCallReport: MetricsReportBase { }
+}
