+++ src/Memores.Metrics.Wcf/Reporters/Counters/Base/RatesCounterBase.cs
@@ -2,8 +2,9 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Memores.Metrics.Wcf.Model;
+using Memores.Metrics.Wcf.Model.Reports;
 
-namespace Memores.Metrics.Wcf.Reporters.Counters {
+namespace Memores.Metrics.Wcf.Reporters.Counters.Base {
     public abstract class RatesCounterBase : ICounter {
         protected readonly IMetricsReporter _reporter;
         private CancellationTokenSource _cancellationTokenSource;
@@ -22,7 +23,7 @@ namespace Memores.Metrics.Wcf.Reporters.Counters {
                     var rate5m = GetRate(currentDateTime, 5);
                     var rate15m = GetRate(currentDateTime, 15);
 
-                    _reporter.Report(new MetricsReport() {
+                    _reporter.Report(new RatesReport() {
                         MetricsReportType = MetricsReportTypes.Rates,
                         Rate1m = rate1m,
                         Rate5m = rate5m,
