+++ src/Memores.Metrics.Wcf/Reporters/Counters/ElasticSearchApdexCounter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Memores.Metrics.Wcf.Model;
+using Memores.Metrics.Wcf.Model.Reports;
 using Memores.Metrics.Wcf.Reporters.Counters.Base;
 
 namespace Memores.Metrics.Wcf.Reporters.Counters {
@@ -10,7 +11,7 @@ namespace Memores.Metrics.Wcf.Reporters.Counters {
         {
             var client = ((ElasticSearchMetricsReporter) _reporter).GetClient();
 
-            var calls = client.Count<MetricsReport>(c => c
+            var calls = client.Count<OperationReport>(c => c
                 .Query(q =>
                     q.Match(m => m.Field(f => f.MetricsReportType).Query(((int) MetricsReportTypes.Operation).ToString())) &&
                     q.DateRange(
@@ -19,7 +20,7 @@ namespace Memores.Metrics.Wcf.Reporters.Counters {
                             .LessThan(currentDateTime)
                     ))).Count;
 
-            var satisfiedCalls = client.Count<MetricsReport>(c => c
+            var satisfiedCalls = client.Count<OperationReport>(c => c
                 .Query(q =>
                     q.Range(m => m.Field(f => f.ProcessingTime).LessThan(threshold)) &&
                     q.Match(m => m.Field(f => f.MetricsReportType).Query(((int) MetricsReportTypes.Operation).ToString())) &&
@@ -29,7 +30,7 @@ namespace Memores.Metrics.Wcf.Reporters.Counters {
                             .LessThan(currentDateTime)
                     ))).Count;
 
-            var toleratedCalls = client.Count<MetricsReport>(c => c
+            var toleratedCalls = client.Count<OperationReport>(c => c
                 .Query(q =>
                     q.Range(m => m.Field(f => f.ProcessingTime).LessThan(threshold*4)) &&
                     q.Range(m => m.Field(f => f.ProcessingTime).GreaterThanOrEquals(threshold)) &&
